requirements.txt
requests==2.32.3
urllib3==2.2.2
python-dotenv==1.0.1
pandas==2.3.3